--- frontend/src/components/cookies.ts
@@ -1,4 +1,0 @@
-import { createCookie } from "react-router";
-
-export const accessTokenCookieManager = createCookie("accessToken");
-export const refreshTokenCookieManager = createCookie("refreshToken");
